Update readme.md
Changed formatting in message templates subsection to reflect updates to other vignettes, fixed a spelling error or two, altered formatting of moderators and preconditions sections for readability.

Note: Not sure if we want to fold in the formatting changes I developed for the message template section into this vignette (since it is essentially the vignette template), but thought I would do the work now and leave it here in a separate branch anyway. If we do like the format, great! Otherwise I can change the other vignettes to match this one.